tech(high-level-tests): configure typescript
