+++ high-level-tests/e2e/package.json
@@ -41,7 +41,8 @@
     "cypress-visual-regression": "^1.5.7",
     "eslint": "^7.20.0",
     "eslint-plugin-cypress": "^2.11.2",
-    "npm-run-all": "4.1.5"
+    "npm-run-all": "4.1.5",
+    "typescript": "^4.7.4"
   },
   "dependencies": {
     "jsonwebtoken": "8.5.1"
